removed MedicationDetailsProvider from MedicationStatusIcon
